requirements.txt
python-dotenv==1.0.0
python-binance==1.0.19
pandas==2.0.3
python-dateutil==2.8.2
requests==2.31.0
websocket-client==1.6.1
numpy==1.24.3
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
